+++ web/common/SubmitCancelButtons.js
